Adds barebones mode to lightning trainer config for the property predictor
Sets barebones=True in property predictor trainer kwargs to minimize logging and speed up training. These additional features are useful in training but completely unnecessary in the property predictor setting.

In the property predictor setting, the **only** reason for why we're using the PL trainer is to natively supported distributed inference. We don't care about logging or any of the other Lightning features.