Update pyproject to be in PEP 621 format.
Previously, we used Poetry to manage our dependencies and build process. This doesn't work well with PyTorch as it has some difficult-to-install dependencies. For now, I've just changed this to be a standard setuptools-based pyproject.toml file. We can revisit this in the future once we have a better solution.